# docs/swagger.yaml
basePath: /api/v1
definitions:
  app.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  handler.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.RiskListResponse:
    properties:
      ip:
        items:
          type: string
        type: array
      user:
        items:
          type: string
        type: array
    type: object
...
  handler.adminCouponCreateReq:
    properties:
      amount_cents:
        type: integer
      description:
        type: string
      discount_rate:
        type: integer
      min_spend_cents:
        type: integer
      price_cents:
        type: integer
      purchasable:
        type: boolean
      status:
        type: string
      title:
        type: string
      type:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - title
    - type
    - valid_from
    - valid_to
    type: object
...
  handler.riskEntryReq:
    properties:
      type:
        type: string
      value:
        type: string
    required:
    - type
    - value
    type: object
  model.Coupon:
    properties:
      amount_cents:
        description: 满减金额（分）
        type: integer
      description:
        type: string
      discount_rate:
        description: 折扣百分比，90 表示 9 折
        type: integer
      id:
        type: integer
      min_spend_cents:
        description: 使用门槛（分）
        type: integer
      price_cents:
        description: 购买价格（分）
        type: integer
      purchasable:
        description: 是否可购买
        type: boolean
      status:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/model.CouponType'
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
...
info:
  contact: {}
  description: SneakerFlash 球鞋秒杀系统接口文档
  title: SneakerFlash API
  version: "1.0"
paths:
  /admin/coupons:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.PageData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 管理台优惠券列表
      tags:
      - 管理后台
    post:
      consumes:
      - application/json
      parameters:
      - description: 优惠券模板
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.adminCouponCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Coupon'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 创建优惠券模板
      tags:
      - 管理后台
  /admin/coupons/{id}:
    delete:
      parameters:
      - description: 优惠券ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 删除优惠券模板
      tags:
      - 管理后台
    put:
      consumes:
      - application/json
      parameters:
      - description: 优惠券ID
        in: path
        name: id
        required: true
        type: integer
      - description: 优惠券模板补丁
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.adminCouponUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Coupon'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 更新优惠券模板
      tags:
      - 管理后台
  /admin/orders:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 订单状态：0未支付 1已支付 2失败 3已取消
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.PageData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 管理台订单列表
      tags:
      - 管理后台
  /admin/audit:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 操作者用户名
        in: query
        name: actor_name
        type: string
      - description: 资源类型
        in: query
        name: resource
        type: string
      - description: 动作
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.PageData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 审计日志列表
      tags:
      - 管理后台
  /admin/products:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.PageData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 管理台商品列表
      tags:
      - 管理后台
  /admin/risk/blacklist:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 名单项
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.riskEntryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 删除黑名单
      tags:
      - 管理后台
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.RiskListResponse'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 查询黑名单
      tags:
      - 管理后台
    post:
      consumes:
      - application/json
      parameters:
      - description: 名单项
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.riskEntryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 新增黑名单
      tags:
      - 管理后台
  /admin/risk/graylist:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 名单项
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.riskEntryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 删除灰名单
      tags:
      - 管理后台
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.RiskListResponse'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 查询灰名单
      tags:
      - 管理后台
    post:
      consumes:
      - application/json
      parameters:
      - description: 名单项
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.riskEntryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 新增灰名单
      tags:
      - 管理后台
  /admin/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AdminStats'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 管理台统计
      tags:
      - 管理后台
  /admin/users:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.PageData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 需要管理员权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 管理台用户列表
      tags:
      - 管理后台
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"